Added chart.py via upload
This is the Python script which produces a time-series chart or graph of a cryptocurrency defined in it's code. The script is being adapted to work and present the graph within the Telegram Messaging App. On its own it only produced an image which get saved into a local filepath of where the script runs from.